@@ -6,6 +6,7 @@
   "license": "GPL-3.0",
   "private": true,
   "dependencies": {
+    "@kubernetes/client-node": "^0.10.2",
     "axios": "0.16.2",
     "bluebird": "3.5.0",
     "body-parser": "1.17.2",
@@ -17,7 +18,6 @@
     "fs-extra-promise": "1.0.1",
     "js-yaml": "3.9.1",
     "jwks-rsa": "1.2.1",
-    "kube-watch": "0.5.4",
     "lodash": "4.17.4",
     "mongoose": "5.0.10",
     "mustache": "2.3.0",
@@ -30,9 +30,9 @@
     "babel-plugin-transform-object-rest-spread": "6.26.0",
     "babel-preset-es2015": "6.24.1",
     "concurrently": "3.5.1",
-    "eslint": "4.5.0",
-    "eslint-config-airbnb-base": "11.3.1",
-    "eslint-plugin-import": "2.7.0",
+    "eslint": "^5.16.0",
+    "eslint-config-airbnb-base": "^13.2.0",
+    "eslint-plugin-import": "^2.18.2",
     "jest": "20.0.4",
     "node-mocks-http": "1.6.4",
     "nodemon": "1.11.0",
